Rename sharing.py to asset_shares.py, add asset_type and sharing_include to enums, address comments, rename methods, clean up unused imports
