+++ smartsheet/models/enums/asset_type.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+class AssetType(str, Enum):
+    """Defines the asset types supported by the Sharing API."""
+    SHEET = 'sheet'
+    REPORT = 'report'
+    SIGHT = 'sight'
+    WORKSPACE = 'workspace'
+    COLLECTION = 'collection'
+    FILE = 'file'
