+++ smartsheet/models/enums/sharing_include.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class SharingInclude(str, Enum):
+    ITEM = 'ITEM',
+    WORKSPACE = 'WORKSPACE'
